feat: implement `POST /api/admin/recipes/reset-data`
This change adds a new service method which can be used to reset all
recipes back to a well-known state.

The path to the backup directory can be spcified with via the
configuration key `Data:BackupDirectoryPath`.

This is done by purging the contents
of the directory specified by the configuration
key `Data:RecipeDirectoryPath` and then copying the files from
`Data:BackupDirectoryPath` into purged directory.